Move some social queries in the proper places. Rename value -> metricValue
