+++ lib/sanbase_web/graphql/resolvers/tech_indicators_resolver.ex
@@ -29,92 +29,6 @@ defmodule SanbaseWeb.Graphql.Resolvers.TechIndicatorsResolver do
     )
   end
 
-  def twitter_mention_count(
-        _root,
-        %{
-          ticker: ticker,
-          from: from,
-          to: to,
-          interval: interval,
-          result_size_tail: result_size_tail
-        },
-        _resolution
-      ) do
-    TechIndicators.twitter_mention_count(
-      ticker,
-      from,
-      to,
-      interval,
-      result_size_tail
-    )
-  end
-
-  def emojis_sentiment(
-        _root,
-        %{
-          from: from,
-          to: to,
-          interval: interval,
-          result_size_tail: result_size_tail
-        },
-        _resolution
-      ) do
-    TechIndicators.emojis_sentiment(
-      from,
-      to,
-      interval,
-      result_size_tail
-    )
-  end
-
-  def social_volume(
-        _root,
-        %{
-          slug: slug,
-          from: from,
-          to: to,
-          interval: interval,
-          social_volume_type: social_volume_type
-        },
-        _resolution
-      ) do
-    TechIndicators.social_volume(
-      slug,
-      from,
-      to,
-      interval,
-      social_volume_type
-    )
-  end
-
-  def social_volume_projects(
-        _root,
-        %{},
-        _resolution
-      ) do
-    TechIndicators.social_volume_projects()
-  end
-
-  def topic_search(
-        _root,
-        %{
-          source: source,
-          search_text: search_text,
-          from: from,
-          to: to,
-          interval: interval
-        },
-        _resolution
-      ) do
-    TechIndicators.topic_search(
-      source,
-      search_text,
-      from,
-      to,
-      interval
-    )
-  end
-
   def metric_anomaly(
         _root,
         %{
