+++ lib/sanbase_web/graphql/schema/types/social_data_types.ex
@@ -33,6 +33,38 @@ defmodule SanbaseWeb.Graphql.SocialDataTypes do
     value(:discord_discussion_overview)
   end
 
+  enum :topic_search_sources do
+    value(:telegram)
+    value(:professional_traders_chat)
+    value(:reddit)
+    value(:discord)
+  end
+
+  object :twitter_mention_count do
+    field(:datetime, non_null(:datetime))
+    field(:mention_count, :integer)
+  end
+
+  object :emojis_sentiment do
+    field(:datetime, non_null(:datetime))
+    field(:sentiment, :float)
+  end
+
+  object :topic_search do
+    field(:messages, list_of(:messages))
+    field(:chart_data, list_of(:chart_data))
+  end
+
+  object :messages do
+    field(:text, :string)
+    field(:datetime, non_null(:datetime))
+  end
+
+  object :chart_data do
+    field(:mentions_count, :integer)
+    field(:datetime, non_null(:datetime))
+  end
+
   object :social_volume do
     field(:datetime, non_null(:datetime))
     field(:mentions_count, :integer)
